+++ packages/kiarina-lib-openai/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Made `api_key` field optional (can be `None`) to support environment variable configuration without explicit settings
+
 ## [1.11.0] - 2025-12-01
 
 ### Added